requirements.txt
# ──────────────────────────────────────────────
# ShelfWatch — Python Dependencies
# ──────────────────────────────────────────────

# Core ML
ultralytics>=8.1.0
torch>=2.1.0
torchvision>=0.16.0

# Inference API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.24.0

# Experiment tracking
mlflow>=2.10.0

# Data download
roboflow>=1.1.0

# Linting & testing
ruff>=0.3.0
pytest>=8.0.0
httpx>=0.27.0          # for testing FastAPI with TestClient

# Observability
prometheus-client>=0.20.0

# Optimized inference
onnxruntime>=1.17.0

# HTTP client (for demo scripts)
requests>=2.31.0
